requirements.txt
requests
datetime
pandas
nltk
numpy
psaw
progressbar2
plotly
dash